Adding `content_root` config option
I'd like to be able to pull in post content from a directory outside of
the site root.  I've added a `content_root` option in order to make that
possible.

Specifying an absolute path to a directory will parse files in that
directory exactly as though they were located in your site's `_posts`
directory, publishing them according to the same rules you've defined in
your site's `_config.yaml`.